Project structure updated
